MOSIP-41291: Sonar duplication fix (#1329)
