+++ pmp-revamp-ui/src/pages/admin/deviceProviderServices/AdminDeviceDetailsFilter.js
@@ -2,7 +2,7 @@ import { useState, useEffect } from "react";
 import DropdownComponent from "../../common/fields/DropdownComponent.js";
 import TextInputComponent from "../../common/fields/TextInputComponent.js";
 import { useTranslation } from "react-i18next";
-import { isLangRTL, createDropdownData, fetchDeviceTypeDropdownData, fetchDeviceSubTypeDropdownData, validateInputRegex } from "../../../utils/AppUtils.js";
+import { isLangRTL, createDropdownData, fetchDeviceTypeDropdownData, fetchDeviceSubTypeDropdownData, validateInputRegex, getFilterDropdownStyle, getFilterTextFieldStyle } from "../../../utils/AppUtils.js";
 import { getUserProfile } from '../../../services/UserProfileService.js';
 import PropTypes from 'prop-types';
 
@@ -98,16 +98,6 @@ function AdminDeviceDetailsFilter({ onApplyFilter, setErrorCode, setErrorMsg, re
             || invalidMake || invalidModel;
     };
 
-    const styles = {
-        dropdownButton: "min-w-64",
-    };
-
-    const styleSet = {
-        inputField: "min-w-64",
-        inputLabel: "mb-2",
-        outerDiv: "ml-4"
-    };
-
     return (
         <div className="flex w-full p-3 justify-start bg-[#F7F7F7] flex-wrap">
             <TextInputComponent
@@ -115,7 +105,7 @@ function AdminDeviceDetailsFilter({ onApplyFilter, setErrorCode, setErrorMsg, re
                 onTextChange={onFilterChangeEvent}
                 fieldNameKey="oidcClientsList.partnerId"
                 placeHolderKey="partnerList.searchPartnerId"
-                styleSet={styleSet}
+                styleSet={getFilterTextFieldStyle()}
                 id="partner_id_filter"
                 inputError={invalidPartnerId}
             />
@@ -124,7 +114,7 @@ function AdminDeviceDetailsFilter({ onApplyFilter, setErrorCode, setErrorMsg, re
                 onTextChange={onFilterChangeEvent}
                 fieldNameKey="oidcClientsList.orgName"
                 placeHolderKey="partnerList.searchOrganisation"
-                styleSet={styleSet}
+                styleSet={getFilterTextFieldStyle()}
                 id="org_name_filter"
                 inputError={invalidOrgName}
             />
@@ -135,7 +125,7 @@ function AdminDeviceDetailsFilter({ onApplyFilter, setErrorCode, setErrorMsg, re
                         onTextChange={onFilterChangeEvent}
                         fieldNameKey="sbiList.sbiId"
                         placeHolderKey="sbiList.searchSbiId"
-                        styleSet={styleSet}
+                        styleSet={getFilterTextFieldStyle()}
                         id="sbi_id_filter"
                         inputError={invalidSbiId}
                     />
@@ -144,7 +134,7 @@ function AdminDeviceDetailsFilter({ onApplyFilter, setErrorCode, setErrorMsg, re
                         onTextChange={onFilterChangeEvent}
                         fieldNameKey="sbiList.sbiVersion"
                         placeHolderKey="sbiList.searchVersion"
-                        styleSet={styleSet}
+                        styleSet={getFilterTextFieldStyle()}
                         id="sbi_version_filter"
                         inputError={invalidSbiVersion}
                     />
@@ -155,7 +145,7 @@ function AdminDeviceDetailsFilter({ onApplyFilter, setErrorCode, setErrorMsg, re
                 onTextChange={onFilterChangeEvent}
                 fieldNameKey="devicesList.deviceId"
                 placeHolderKey="devicesList.searchDeviceId"
-                styleSet={styleSet}
+                styleSet={getFilterTextFieldStyle()}
                 id="device_id_filter"
                 inputError={invalidDeviceId}
             />
@@ -166,7 +156,7 @@ function AdminDeviceDetailsFilter({ onApplyFilter, setErrorCode, setErrorMsg, re
                 fieldNameKey='addDevices.deviceType'
                 placeHolderKey='addDevices.selectDeviceType'
                 isPlaceHolderPresent={true}
-                styleSet={styles}
+                styleSet={getFilterDropdownStyle()}
                 id='device_type_filter'>
             </DropdownComponent>
             <DropdownComponent
@@ -175,7 +165,7 @@ function AdminDeviceDetailsFilter({ onApplyFilter, setErrorCode, setErrorMsg, re
                 onDropDownChangeEvent={onFilterChangeEvent}
                 fieldNameKey='addDevices.deviceSubType'
                 placeHolderKey='addDevices.selectDeviceSubType'
-                styleSet={styles}
+                styleSet={getFilterDropdownStyle()}
                 disabled={filters.deviceType === ""}
                 isPlaceHolderPresent={true}
                 id='device_sub_type_filter'>
@@ -185,7 +175,7 @@ function AdminDeviceDetailsFilter({ onApplyFilter, setErrorCode, setErrorMsg, re
                 onTextChange={onFilterChangeEvent}
                 fieldNameKey="ftmList.make"
                 placeHolderKey="ftmList.searchMake"
-                styleSet={styleSet}
+                styleSet={getFilterTextFieldStyle()}
                 id="make_filter"
                 inputError={invalidMake}
             />
@@ -194,7 +184,7 @@ function AdminDeviceDetailsFilter({ onApplyFilter, setErrorCode, setErrorMsg, re
                 onTextChange={onFilterChangeEvent}
                 fieldNameKey="ftmList.model"
                 placeHolderKey="ftmList.searchModel"
-                styleSet={styleSet}
+                styleSet={getFilterTextFieldStyle()}
                 id="model_filter"
                 inputError={invalidModel}
             />
@@ -204,7 +194,7 @@ function AdminDeviceDetailsFilter({ onApplyFilter, setErrorCode, setErrorMsg, re
                 onDropDownChangeEvent={onFilterChangeEvent}
                 fieldNameKey="partnerList.status"
                 placeHolderKey="partnerList.selectStatus"
-                styleSet={styles}
+                styleSet={getFilterDropdownStyle()}
                 isPlaceHolderPresent={true}
                 id="status_filter"
             />
