+++ pmp-revamp-ui/src/pages/admin/policyManager/PolicyGroupListFilter.js
@@ -2,7 +2,7 @@ import { useState, useEffect } from "react";
 import DropdownComponent from "../../common/fields/DropdownComponent.js";
 import TextInputComponent from "../../common/fields/TextInputComponent.js";
 import { useTranslation } from "react-i18next";
-import { isLangRTL, createDropdownData, validateInputRegex } from "../../../utils/AppUtils.js";
+import { isLangRTL, createDropdownData, validateInputRegex, getFilterDropdownStyle, getFilterTextFieldStyle } from "../../../utils/AppUtils.js";
 import { getUserProfile } from '../../../services/UserProfileService';
 import PropTypes from 'prop-types';
 
@@ -47,16 +47,6 @@ function PolicyGroupListFilter({ onApplyFilter }) {
     const areFiltersEmpty = () => {
         return Object.values(filters).every(value => value === "") || invalidPolicyGroupId || invalidPolicyGroupName || invalidPolicyGroupDescr;
       };
-    
-    const styles = {
-      dropdownButton: "min-w-64",
-    };
-    
-    const styleSet = {
-      inputField: "min-w-64",
-      inputLabel: "mb-2",
-      outerDiv: "ml-4"
-    };
 
     return (
         <div className="flex w-full p-2.5 justify-start bg-[#F7F7F7] flex-wrap">
@@ -65,7 +55,7 @@ function PolicyGroupListFilter({ onApplyFilter }) {
             onTextChange={onFilterChangeEvent}
             fieldNameKey="policyGroupList.policyGroupId"
             placeHolderKey="policyGroupList.searchPolicyGroupId"
-            styleSet={styleSet}
+            styleSet={getFilterTextFieldStyle()}
             id="policy_group_id_filter"
             inputError={invalidPolicyGroupId}
           />
@@ -74,7 +64,7 @@ function PolicyGroupListFilter({ onApplyFilter }) {
             onTextChange={onFilterChangeEvent}
             fieldNameKey="policyGroupList.policyGroupName"
             placeHolderKey="policyGroupList.searchPolicyGroupName"
-            styleSet={styleSet}
+            styleSet={getFilterTextFieldStyle()}
             id="policy_group_name_filter"
             inputError={invalidPolicyGroupName}
           />
@@ -83,7 +73,7 @@ function PolicyGroupListFilter({ onApplyFilter }) {
             onTextChange={onFilterChangeEvent}
             fieldNameKey="policyGroupList.policyGroupDescription"
             placeHolderKey="policyGroupList.searchPolicyGroupDescription"
-            styleSet={styleSet}
+            styleSet={getFilterTextFieldStyle()}
             id="policy_group_description_filter"
             inputError={invalidPolicyGroupDescr}
           />
@@ -93,7 +83,7 @@ function PolicyGroupListFilter({ onApplyFilter }) {
             onDropDownChangeEvent={onFilterChangeEvent}
             fieldNameKey="partnerList.status"
             placeHolderKey="partnerList.selectStatus"
-            styleSet={styles}
+            styleSet={getFilterDropdownStyle()}
             isPlaceHolderPresent={true}
             id="status_filter"
           />
